+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+#### 0.9.5
+* `Expression` support for `OPTION` value
+
+#### 0.9.4
+* Replaced `getConnectionParams()` and `setConnectionParams()` with `getParam()`, `getParams()`, `setParam()` (thanks to @FindTheBest)
+* Deprecated `getConnectionParams()` and `setConnectionParams()`
+* Added `ConnectionInterface`
+
 #### 0.9.3
 
 * HHVM support